Update to version 5.0
New Advanced Permissions System – Unlimited roles types
Super Administrators
Administrators
Manager
Principal
Teacher
Student
Parents
Accountant
Librarian
HR
Receptionist
Store Keeper
Front Office
Visitors Log
Track your School visitors
Track purpose of visits
Track Check-in & Check-Out
Phone Calls
Track your incoming/outgoing phone calls
Track purpose, Type, duration and details of calls
Postal
Track your dispatch/received postal
Contact Messages: See your website visitor messages directly from Web Portal
Front Office Categories to easily manage sources, types & Purposes
Enquiries
Track classified Enquiries from Parents, Students, Teachers,…
Complaints
Track classified Complaints from Parents, Students, Teachers,…
More languages supported, You still can create your own language if your language isn’t in the list.
English
Arabic
French
Dutch
German
Hindi
Italian
Turkish
Russian
Spanish
Portuguse
Bengali
Chineese
Indonesian
Romanian
Thai
Manage Employees
Mange departments & designations
Enroll your school employees in the system
Classify each employee under dept. & design.
Easily control each employee roles & permissions
Students
New smart & advanced school admission form
Control student categories
Control Medical History
Enhanced view of father & mother information
Enhanced view of Guardian account information
Control students documents
Student Attendance
Add notes when taking the attendance
Attendance report redesigned completely
One view to see multiple students attendance with period
Calculation of presence percentage for each student
Staff Attendance
Add notes when taking the attendance
Check-In & Check-Out attendance tracking
Attendance report redesigned completely
One view to see multiple students attendance with period
Calculation of presence percentage for each student
Vacation
Completely redesigned vacation system
Approve/Decline Vacation requests
Users able to track their approved & declined vacation requests
Library Management
Completely redesigned library system
Mange users subscription to library system including employees and students
Issue/Return books to users
Track issued books to users, Dates and return dates
Inventory Management
Manage Stores
Manage Suppliers
Manage Inventory Categories
Manage items coding
Manage Items instock
Manage issued items
Manage returned items
Online Exams
Questions Bank
Randomize exam questions supported
Sharing questions, To allow question to be used by another teacher in another exam
Rich text editor for questions
Can attach images for each question
Unlimited number of questions supported
Unlimited number of options per each question supported
New UI for online exam taking
FrontEnd
New modern frontend design
Frontend permalink enhancements
Frontend pages order list
Dashboard
Today’s Birthday widget
Dashboard Widgets rearrangements
Enhanced view of exam schedule
Calculating total income
Calculating total expenses
Clickatell SMS Gateway support new API system
Salary Delete Payroll feature

+++ app/Http/Controllers/NewsBoardController.php
@@ -8,7 +8,7 @@ class NewsBoardController extends Controller {
 	var $layout = 'dashboard';
 
 	public function __construct(){
-		if(app('request')->header('Authorization') != ""){
+		if(app('request')->header('Authorization') != "" || \Input::has('token')){
 			$this->middleware('jwt.auth');
 		}else{
 			$this->middleware('authApplication');
@@ -22,19 +22,21 @@ class NewsBoardController extends Controller {
 			return \Redirect::to('/');
 		}
 
-		if(!$this->panelInit->hasThePerm('newsboard')){
-			exit;
-		}
 	}
 
 	public function listAll($page = 1)
 	{
+
+		if(!$this->panelInit->can( array("newsboard.list","newsboard.View","newsboard.addNews","newsboard.editNews","newsboard.delNews") )){
+			exit;
+		}
+
 		$toReturn = array();
 		if($this->data['users']->role == "admin" ){
-			$toReturn['newsboard'] = \newsboard::orderby('id','DESC')->take('20')->skip(20* ($page - 1) )->get()->toArray();
+			$toReturn['newsboard'] = \newsboard::orderby('newsDate','DESC')->take('20')->skip(20* ($page - 1) )->get()->toArray();
 			$toReturn['totalItems'] = \newsboard::count();
 		}else{
-			 $toReturn['newsboard'] = \newsboard::where('newsFor',$this->data['users']->role)->orWhere('newsFor','all')->orderby('id','DESC')->take('20')->skip(20* ($page - 1) )->get()->toArray();
+			 $toReturn['newsboard'] = \newsboard::where('newsFor',$this->data['users']->role)->orWhere('newsFor','all')->orderby('newsDate','DESC')->take('20')->skip(20* ($page - 1) )->get()->toArray();
 			 $toReturn['totalItems'] = \newsboard::where('newsFor',$this->data['users']->role)->orWhere('newsFor','all')->count();
 		}
 
@@ -43,13 +45,16 @@ class NewsBoardController extends Controller {
 			$toReturn['newsboard'][$key]['newsDate'] = $this->panelInit->unix_to_date($toReturn['newsboard'][$key]['newsDate']);
 		}
 
-		$toReturn['userRole'] = $this->data['users']->role;
-
 		return $toReturn;
 	}
 
 	public function search($keyword,$page = 1)
 	{
+		
+		if(!$this->panelInit->can( array("newsboard.View","newsboard.addNews","newsboard.editNews","newsboard.delNews") )){
+			exit;
+		}
+
 		$toReturn = array();
 		if($this->data['users']->role == "admin" ){
 			$toReturn['newsboard'] = \newsboard::where('newsTitle','like','%'.$keyword.'%')->orWhere('newsText','like','%'.$keyword.'%')->take('20')->skip(20* ($page - 1) )->get()->toArray();
@@ -64,12 +69,15 @@ class NewsBoardController extends Controller {
 			$toReturn['newsboard'][$key]['newsDate'] = $this->panelInit->unix_to_date($toReturn['newsboard'][$key]['newsDate']);
 		}
 
-		$toReturn['userRole'] = $this->data['users']->role;
-
 		return $toReturn;
 	}
 
 	public function delete($id){
+
+		if(!$this->panelInit->can( "newsboard.delNews" )){
+			exit;
+		}
+
 		if ( $postDelete = \newsboard::where('id', $id)->first() )
         {
             $postDelete->delete();
@@ -80,7 +88,11 @@ class NewsBoardController extends Controller {
 	}
 
 	public function create(){
-		if($this->data['users']->role != "admin") exit;
+		
+		if(!$this->panelInit->can( "newsboard.addNews" )){
+			exit;
+		}
+
 		$newsboard = new \newsboard();
 		$newsboard->newsTitle = \Input::get('newsTitle');
 		$newsboard->newsText = htmlspecialchars(\Input::get('newsText'),ENT_QUOTES);
@@ -91,6 +103,11 @@ class NewsBoardController extends Controller {
 
 		if (\Input::hasFile('newsImage')) {
 			$fileInstance = \Input::file('newsImage');
+
+			if(!$this->panelInit->validate_upload($fileInstance)){
+				return $this->panelInit->apiOutput(false,$this->panelInit->language['addNews'],"Sorry, This File Type Is Not Permitted For Security Reasons ");
+			}
+
 			$newFileName = uniqid().".".$fileInstance->getClientOriginalExtension();
 			$fileInstance->move('uploads/news/',$newFileName);
 
@@ -123,6 +140,11 @@ class NewsBoardController extends Controller {
 	}
 
 	function fetch($id){
+
+		if(!$this->panelInit->can( array("newsboard.View","newsboard.editNews") )){
+			exit;
+		}
+
 		$data = \newsboard::where('id',$id)->first()->toArray();
 		$data['newsText'] = htmlspecialchars_decode($data['newsText'],ENT_QUOTES);
 		$data['newsDate'] = $this->panelInit->unix_to_date($data['newsDate']);
@@ -130,7 +152,11 @@ class NewsBoardController extends Controller {
 	}
 
 	function edit($id){
-		if($this->data['users']->role != "admin") exit;
+
+		if(!$this->panelInit->can( "newsboard.editNews" )){
+			exit;
+		}
+
 		$newsboard = \newsboard::find($id);
 		$newsboard->newsTitle = \Input::get('newsTitle');
 		$newsboard->newsText = htmlspecialchars(\Input::get('newsText'),ENT_QUOTES);
@@ -140,6 +166,11 @@ class NewsBoardController extends Controller {
 
 		if (\Input::hasFile('newsImage')) {
 			$fileInstance = \Input::file('newsImage');
+
+			if(!$this->panelInit->validate_upload($fileInstance)){
+				return $this->panelInit->apiOutput(false,$this->panelInit->language['editNews'],"Sorry, This File Type Is Not Permitted For Security Reasons ");
+			}
+			
 			$newFileName = uniqid().".".$fileInstance->getClientOriginalExtension();
 			$fileInstance->move('uploads/news/',$newFileName);
 
@@ -154,7 +185,11 @@ class NewsBoardController extends Controller {
 	}
 
 	function fe_active($id){
-		if($this->data['users']->role != "admin") exit;
+		
+		if(!$this->panelInit->can( "newsboard.editNews" )){
+			exit;
+		}
+		
 		$newsboard = \newsboard::find($id);
 		
 		if($newsboard->fe_active == 1){
